+++ src/Fantomas.Core/SourceParser.fs
@@ -1071,7 +1071,8 @@ let (|Record|_|) =
 
 let (|AnonRecord|_|) =
     function
-    | SynExpr.AnonRecd (isStruct, copyInfo, fields, _) -> Some(isStruct, fields, Option.map fst copyInfo)
+    | SynExpr.AnonRecd (isStruct, copyInfo, fields, StartEndRange 1 (openingBrace, _, closingBrace)) ->
+        Some(openingBrace, isStruct, fields, Option.map fst copyInfo, closingBrace)
     | _ -> None
 
 let (|ObjExpr|_|) =