+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [5.1.0-beta-002] - 2022-10-29
 
 ### Fixed
 * Idempotency problem with `struct end`. [#2592](https://github.com/fsprojects/fantomas/issues/2592)
@@ -11,7 +11,7 @@
 ## [5.1.0-beta-001] - 2022-10-19
 
 ### Fixed
-* Comments get removed for method parameter with attribute [#2585](https://github.com/fsprojects/fantomas/issues/2585)
+* Comments get removed for method parameter with attribute. [#2585](https://github.com/fsprojects/fantomas/issues/2585)
 
 ### Changed
 * Update FCS to 'TypeDefn and NamespaceOrModule leading trivia', commit 7d115d56cde12006dd26a3c5babd702df451b8e8